+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.3.0] - 2021-05-09
+### Changed
+- `onlyAvailable` filter is now `onlyWithBikes` to be coherent with `onlyWithParking`
+- `bikesnumber` field is now `bikescount` to be coherent with `emptyslotcount`
+- `slotsnumber` field is now `slotscount` to be coherent with `emptyslotcount`
+- `emptyslotcount` field is now `emptyslotscount` to be coherent with `slotsnumber` and `bikesnumber`
 
 ## [5.2.0] - 2021-05-09
 ### Added